requirements.txt
argparse
easydict
h5py
matplotlib
numpy==1.26.4
open3d==0.18.0
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5
tqdm
transformers==4.45.2
transforms3d
termcolor
tensorboard
faiss-cpu
torch_cluster
plyfile